Created user routers
